code tweaks for clean compile 20+ years later

+++ SelFile/Draw.c
@@ -491,7 +491,7 @@ SFscrollTimer(p, id)
 	int	save;
 	int     n;
 
-        n = (int) p;
+        n = (intptr_t) p;
 
 	dir = &(SFdirs[SFdirPtr + n]);
 	save = dir->vOrigin;
@@ -521,7 +521,7 @@ SFscrollTimer(p, id)
 
 	if (SFbuttonPressed) {
 		SFscrollTimerId = XtAppAddTimeOut(SFapp,
-			SFscrollTimerInterval(), SFscrollTimer, (XtPointer) n);
+			SFscrollTimerInterval(), SFscrollTimer, (XtPointer)(intptr_t) n);
 	}
 }
 
@@ -562,7 +562,7 @@ SFnewInvertEntry(n, event)
 				SFscrollTimerAdded = 1;
 				SFscrollTimerId = XtAppAddTimeOut(SFapp,
 					SFscrollTimerInterval(), SFscrollTimer,
-					(XtPointer) n);
+					(XtPointer)(intptr_t) n);
 			}
 		}
